+++ app/Http/Controllers/EtablissementController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Etablissement;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EtablissementController extends Controller
 {
@@ -12,6 +13,11 @@ class EtablissementController extends Controller
      */
     public function index()
     {
+        // Vérification des autorisations
+        if (!Gate::allows('view', Etablissement::class)) {
+            return response()->json(['message' => 'Non autorisé à voir la liste des établissements.'], 403);
+        }
+        // Récupération de tous les établissements
         $etablissements = Etablissement::with(['directeurEtablissement'])->get();
         return response()->json([
             'message' => 'Liste des établissements récupérée avec succès',
@@ -30,6 +36,11 @@ class EtablissementController extends Controller
             'telephone' => 'required|string',
             'directeur_regional_id' => 'required|exists:users,id'
         ]);
+        // Vérification des autorisations
+        if (!Gate::allows('create', Etablissement::class)) {
+            return response()->json(['message' => 'Non autorisé à créer un établissement.'], 403);
+        }
+        // Création de l'établissement
 
         $etablissement = Etablissement::create($data);
 
@@ -44,6 +55,11 @@ class EtablissementController extends Controller
      */
     public function show(Etablissement $etablissement)
     {
+        // Vérification des autorisations
+        if (!Gate::allows('view', $etablissement)) {
+            return response()->json(['message' => 'Non autorisé à voir cet établissement.'], 403);
+        }
+        // Récupération de l'établissement avec le directeur
         return response()->json([
             'message' => 'Etablissement récupéré avec succès',
             'data' => $etablissement->load(['directeurEtablissement']),
@@ -61,6 +77,11 @@ class EtablissementController extends Controller
             'telephone' => 'sometimes|string',
             'directeur_regional_id' => 'sometimes|exists:users,id'
         ]);
+        // Vérification des autorisations
+        if (!Gate::allows('update', $etablissement)) {
+            return response()->json(['message' => 'Non autorisé à mettre à jour cet établissement.'], 403);
+        }
+        // Mise à jour de l'établissement
 
         $etablissement->update($data);
 
@@ -75,6 +96,10 @@ class EtablissementController extends Controller
      */
     public function destroy(Etablissement $etablissement)
     {
+        // Vérification des autorisations
+        if (!Gate::allows('delete', $etablissement)) {
+            return response()->json(['message' => 'Non autorisé à supprimer cet établissement.'], 403);
+        }
         $etablissement->delete();
 
         return response()->json(null, 204);
